Move merging code to separate file
Easier to unit test all the permutations
